profile viewer stuff

+++ src/main/java/julianh06/wynnextras/utils/overlays/EasyButton.java
@@ -35,7 +35,7 @@ public class EasyButton extends EasyElement{
     }
 
     public void drawWithTexture(DrawContext context, Identifier texture) {
-        RenderUtils.drawTexturedRect(context.getMatrices(), texture, x, y, width, height, width, height);
+        RenderUtils.drawTexturedRect(context.getMatrices(), texture, x, y, width, height, (int) width, (int) height);
         if(buttonText == null) {
             return;
         }
